@@ -1,5 +1,6 @@
 const { contextBridge, ipcRenderer } = require('electron');
 
 contextBridge.exposeInMainWorld('api', {
-  getMovies: () => ipcRenderer.invoke('get-movies')
+  getUnavailableSeats: (movieId) => ipcRenderer.invoke('get-unavailable-seats', movieId),
+  bookSeat: (movieId, rowNumber, seatNumber) => ipcRenderer.invoke('book-seat', movieId, rowNumber, seatNumber)
 });
